api/requirements.txt
psycopg[binary]>=3.2.8,<3.3
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
alembic==1.13.2
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic_settings
bcrypt==3.2.2